refactor: fixed type imports
