+++ assets/ArrowUpIcon.tsx
@@ -1,5 +1,5 @@
 //? Import IconProperties for props typecasting
-import { IconProperties } from "../types/asset-properties/IconProperties.ts";
+import type { IconProperties } from "../types/asset-properties/IconProperties.ts";
 
 //? Renders an Arrow Up Icon with fill
 export function ArrowUpIcon(
