+++ data/blog/how-create-theme-switcher-deno-fresh.ts
@@ -1,5 +1,5 @@
 //? Import required properties for a blog post
-import BlogPostSummaryProperties from "../../types/blog/BlogPostSummaryProperties.ts";
+import type { BlogPostSummaryProperties } from "../../types/blog/BlogPostSummaryProperties.ts";
 
 //? Export object with blog post summary data
 export const createFreshThemeSwitcher: BlogPostSummaryProperties = {
