+++ services/tictactoe/resetBoardFunction.ts
@@ -1,7 +1,7 @@
 //? Import maximum number of tiles in a board
 import { MAX_TICTACTOE_TILES } from "../../data/projects/tictactoe/maxTiles.ts";
 //? Import interface for typecasting
-import { BoardTile } from "../../types/component-properties/projects/tictactoe/BoardTile.ts";
+import type { BoardTile } from "../../types/component-properties/projects/tictactoe/BoardTile.ts";
 
 //? Reset the board and populate it with 9 empty elements to render new tiles
 export function resetBoardFunction(board: BoardTile[]) {
